requirements.txt
streamlit
openai
langchain_openai
langchain_community
langchain
python-dotenv
langchain_core
langchain_groq